+++ src/components/Countdown.jsx
@@ -1,0 +1,27 @@
+import React, { useEffect, useState } from "react";
+
+const Countdown = ({ expiryDate }) => {
+  const [now, setNow] = useState(Date.now());
+
+  useEffect(() => {
+    const interval = setInterval(() => setNow(Date.now()), 1000);
+    return () => clearInterval(interval);
+  }, []);
+
+  const formatTime = (expiry) => {
+    const diff = expiry - now;
+    if (diff <= 0) return null;
+    const totalSeconds = Math.floor(diff / 1000);
+    const hours = Math.floor(totalSeconds / 3600);
+    const minutes = Math.floor((totalSeconds % 3600) / 60);
+    const seconds = totalSeconds % 60;
+    return `${hours}h ${minutes}m ${seconds}s`;
+  };
+
+  const timeString = formatTime(expiryDate);
+  if (!timeString) return null;
+
+  return <div className="de_countdown">{timeString}</div>;
+};
+
+export default Countdown;
