+++ src/components/home/NewItems.jsx
@@ -5,13 +5,13 @@ import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import Slider from "react-slick";
 import "./HotCollections.css";
+import Countdown from "../Countdown";
 
 const NewItems = () => {
   const [newItems, setNewItems] = useState([]);
   const [loading, setLoading] = useState(false);
   const fullArray = [1, 2, 3, 4];
   const [visibleItems, setVisibleItems] = useState(fullArray);
-  const [now, setNow] = useState(Date.now());
 
   async function getNewItems() {
     setLoading(true);
@@ -40,21 +40,6 @@ const NewItems = () => {
 
     return () => window.removeEventListener("resize", updateVisibleItems);
   }, []);
-
-  useEffect(() => {
-    const interval = setInterval(() => setNow(Date.now()), 1000);
-    return () => clearInterval(interval);
-  }, []);
-
-  const formatTime = (expiryDate) => {
-    const diff = expiryDate - now;
-    if (diff <= 0) return "Expired";
-    const totalSeconds = Math.floor(diff / 1000);
-    const hours = Math.floor(totalSeconds / 3600);
-    const minutes = Math.floor((totalSeconds % 3600) / 60);
-    const seconds = totalSeconds % 60;
-    return `${hours}h ${minutes}m ${seconds}s`;
-  };
 
   var settings = {
     dots: false,
@@ -111,78 +96,71 @@ const NewItems = () => {
           </div>
           {
             <Slider {...settings}>
-              {newItems.map((newItem, index) => {
-                const diff = newItem.expiryDate - now;
-                return (
-                  <div className="carousel-wrapper">
-                    <div className="px-2" key={index}>
-                      <div className="nft__item">
-                        <div className="author_list_pp">
-                          <Link
-                            to={`/author/${newItem.authorId}`}
-                            data-bs-toggle="tooltip"
-                            data-bs-placement="top"
-                            title="Creator: Monica Lucas"
-                          >
-                            <img
-                              className="lazy"
-                              src={newItem.authorImage}
-                              alt=""
-                            />
-                            <i className="fa fa-check"></i>
-                          </Link>
-                        </div>
+              {newItems.map((newItem, index) => (
+                <div className="carousel-wrapper">
+                  <div className="px-2" key={index}>
+                    <div className="nft__item">
+                      <div className="author_list_pp">
+                        <Link
+                          to={`/author/${newItem.authorId}`}
+                          data-bs-toggle="tooltip"
+                          data-bs-placement="top"
+                          title="Creator: Monica Lucas"
+                        >
+                          <img
+                            className="lazy"
+                            src={newItem.authorImage}
+                            alt=""
+                          />
+                          <i className="fa fa-check"></i>
+                        </Link>
+                      </div>
 
-                        {diff > 0 && (
-                          <div className="de_countdown">
-                            {formatTime(newItem.expiryDate)}
-                          </div>
-                        )}
+                      <Countdown expiryDate={newItem.expiryDate} />
 
-                        <div className="nft__item_wrap">
-                          <div className="nft__item_extra">
-                            <div className="nft__item_buttons">
-                              <button>Buy Now</button>
-                              <div className="nft__item_share">
-                                <h4>Share</h4>
-                                <a href="" target="_blank" rel="noreferrer">
-                                  <i className="fa fa-facebook fa-lg"></i>
-                                </a>
-                                <a href="" target="_blank" rel="noreferrer">
-                                  <i className="fa fa-twitter fa-lg"></i>
-                                </a>
-                                <a href="">
-                                  <i className="fa fa-envelope fa-lg"></i>
-                                </a>
-                              </div>
+                      <div className="nft__item_wrap">
+                        <div className="nft__item_extra">
+                          <div className="nft__item_buttons">
+                            <button>Buy Now</button>
+                            <div className="nft__item_share">
+                              <h4>Share</h4>
+                              <a href="" target="_blank" rel="noreferrer">
+                                <i className="fa fa-facebook fa-lg"></i>
+                              </a>
+                              <a href="" target="_blank" rel="noreferrer">
+                                <i className="fa fa-twitter fa-lg"></i>
+                              </a>
+                              <a href="">
+                                <i className="fa fa-envelope fa-lg"></i>
+                              </a>
                             </div>
                           </div>
-
-                          <Link to={`/item-details/${newItem.nftId}`}>
-                            <img
-                              src={newItem.nftImage}
-                              className="lazy nft__item_preview"
-                              alt=""
-                            />
-                          </Link>
                         </div>
-                        <div className="nft__item_info">
-                          <Link to={`/item-details/${newItem.nftId}`}>
-                            <h4>{newItem.title}</h4>
-                          </Link>
-                          <div className="nft__item_price">
-                            {newItem.price} ETH
-                          </div>
-                          <div className="nft__item_like">
-                            <i className="fa fa-heart"></i>
-                            <span>{newItem.likes}</span>
-                          </div>
+
+                        <Link to={`/item-details/${newItem.nftId}`}>
+                          <img
+                            src={newItem.nftImage}
+                            className="lazy nft__item_preview"
+                            alt=""
+                          />
+                        </Link>
+                      </div>
+                      <div className="nft__item_info">
+                        <Link to={`/item-details/${newItem.nftId}`}>
+                          <h4>{newItem.title}</h4>
+                        </Link>
+                        <div className="nft__item_price">
+                          {newItem.price} ETH
+                        </div>
+                        <div className="nft__item_like">
+                          <i className="fa fa-heart"></i>
+                          <span>{newItem.likes}</span>
                         </div>
                       </div>
                     </div>
                   </div>
-                );
-              })}
+                </div>
+              ))}
             </Slider>
           }
           <style>{`.slick-prev:before,
